backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
PyYAML==6.0.1